+++ bindings/src/lib.rs
@@ -17,7 +17,14 @@ pub fn parse_recipe(input: String) -> CooklangRecipe {
     let converter = Converter::empty();
 
     let mut parser = PullParser::new(&input, extensions);
-    let parsed = parse_events(&mut parser, &input, extensions, &converter, None).unwrap_output();
+    let parsed = parse_events(
+        &mut parser,
+        &input,
+        extensions,
+        &converter,
+        Default::default(),
+    )
+    .unwrap_output();
 
     into_simple_recipe(&parsed)
 }
@@ -35,7 +42,7 @@ pub fn parse_metadata(input: String) -> CooklangMetadata {
         &input,
         extensions,
         &converter,
-        None,
+        Default::default(),
     )
     .map(|c| c.metadata.map)
     .unwrap_output();